Add an experiment in creating an ldraw model file
This was an experiment in recreating the steps to make a candy cane I
did years ago.
I wanted to create a round shape and then optimize from 1x1 to larger
parts.
Unfortuntately, it hit a snag in how parts are centered.
But at least the output was a valid ldraw file format.